[#338] Split question setting into component & fix translation setting tab
