+++ frontend/src/components/survey-editor/QuestionEditor.jsx
@@ -19,7 +19,7 @@ import {
   RiEditFill,
   RiTranslate2,
 } from "react-icons/ri";
-import QuestionSetting from "./QuestionSetting";
+import QuestionTabContent from "./QuestionTabContent";
 import { store, api } from "../../lib";
 import { isoLangs } from "../../lib";
 import { useNotification } from "../../util";
@@ -61,6 +61,9 @@ const TranslationTab = ({ activeLang, setActiveLang }) => {
 };
 
 const QuestionMenu = ({ activeSetting, setActiveSetting }) => {
+  const { surveyEditor } = store.useState((s) => s);
+  const { languages } = surveyEditor;
+
   return (
     <Space direction="vertical" size={1} className="question-menu-wrapper">
       <Tooltip title="Show question setting">
@@ -78,6 +81,7 @@ const QuestionMenu = ({ activeSetting, setActiveSetting }) => {
           type="text"
           icon={<RiTranslate2 />}
           onClick={() => setActiveSetting("translation")}
+          disabled={!languages?.length || !languages}
         />
       </Tooltip>
     </Space>
@@ -97,7 +101,7 @@ const QuestionEditor = ({
   toggleMove = false,
 }) => {
   const { surveyEditor, optionValues } = store.useState((s) => s);
-  const { questionGroup: questionGroupState } = surveyEditor;
+  const { languages, questionGroup: questionGroupState } = surveyEditor;
 
   const { question_type } = optionValues;
   const qId = question?.id;
@@ -112,7 +116,19 @@ const QuestionEditor = ({
   const [coreMandatory, setCoreMandatory] = useState(false);
   const [personalData, setPersonalData] = useState(false);
   const [datapointName, setDatapointName] = useState(false);
-  const [activeLang, setActiveLang] = useState(surveyEditor?.languages?.[0]);
+  const [activeLang, setActiveLang] = useState(null);
+
+  // handle when form languages updated
+  useEffect(() => {
+    if (languages?.length) {
+      setActiveLang(languages?.[0]);
+    } else {
+      setActiveLang(null);
+      if (activeSetting === "translation") {
+        setActiveSetting("detail");
+      }
+    }
+  }, [languages, activeSetting]);
 
   useEffect(() => {
     if (qId) {
@@ -348,7 +364,7 @@ const QuestionEditor = ({
                       />
                     </Col>
                     <Col className="input-wrapper">
-                      <QuestionSetting
+                      <QuestionTabContent
                         form={form}
                         activeSetting={activeSetting}
                         questionGroup={questionGroup}
